+++ sample_data/run_calcWB_2023_sample.R
@@ -3,10 +3,9 @@ rm(list=ls())
 try(dev.off())
 
 ### 0.2 configuration ---
-mypath="C:/Users/username/Desktop/waterbalancemodel-main/waterbalancemodel-main/sample_data" #replace example path with your local directory
+mypath="C:/Users/username/Desktop/WaterBalanceMod_Git/waterbalancemodel/sample_data" #replace example path with your local directory
 shape_site = sf::st_read(paste(mypath,"/Shapefile/sample_2023.shp",sep="")) #shapefile of AOI
 target_res=5 #spatial resolution of results, must be >=5
-method_NDVI="uwdw"# uwdw or direct, leave at uwdw
 modeltype="poly" #poly or linear, leave at poly
 last_NDVI_0=132 #last day with NDVI==0, so with no visible vegetation
 ET_ref=read.csv(paste(mypath,"/DWD_ET0_2023.csv",sep=""),sep=",") #csv file with with daily reference ET or leave at NA. If NA, then...
@@ -26,7 +25,6 @@ arable_pass="password" #password for arable account, if you choose to download r
 test_wb=WaterBalanceMod::calcWB(mypath=mypath,
                shape_site=shape_site,
                target_res=target_res,
-               method_NDVI=method_NDVI,
                modeltype=modeltype,
                last_NDVI_0=last_NDVI_0,
                ET_ref=ET_ref,
@@ -44,8 +42,8 @@ test_wb=WaterBalanceMod::calcWB(mypath=mypath,
 
 ############# plot procedure ----
 buffer20 = sf::st_read(paste(mypath,"/Shapefile/Buffer_5_WB.shp",sep=""))
-WaterBalanceMod::calcWBplots(source_path=paste(mypath,"/",method_NDVI,"_",modeltype,"_",precip_source,"_",irrigation_efficiency,"_",last_NDVI_0,"_",target_res,"/",
-                                      "WBM_",method_NDVI,"_",precip_source,"_",modeltype,"_",irrigation_efficiency,"_",last_NDVI_0,"_",target_res,".RData",sep=""),
+WaterBalanceMod::calcWBplots(source_path=paste(mypath,"/",modeltype,"_",precip_source,"_",irrigation_efficiency,"_",last_NDVI_0,"_",target_res,"/",
+                                      "WBM_",precip_source,"_",modeltype,"_",irrigation_efficiency,"_",last_NDVI_0,"_",target_res,".RData",sep=""),
                     plant_doy=109,
                     buffer20=buffer20,
                     shape_site=shape_site)